Add files via upload
Here we have the final codes after multiple trials to reach optimal solution for the news recommendations system. This file includes training, validation, testing and deployment codes however, it is advised to run them alongside with the training dependencies to get use of trained models and saved as well as optimized parameters which is in the training dependencies file.